Changed GetOrigin() in cxIGTLinkSonixStatusMessage from double to float
Former-commit-id: e45b0316994639d74367e3113a056c32cf0cf202

+++ modules/grabberCommon/cxIGTLinkSonixStatusMessage.h
@@ -63,7 +63,7 @@ public:
 
   void SetOrigin(double o[3]);
   void SetOrigin(double oi, double oj, double ok);
-  void GetOrigin(double o[3]);
+  void GetOrigin(float o[3]);
   void GetOrigin(double &oi, double &oj, double &ok);
   void SetSpacing(double si, double sj, double sk);
   void GetSpacing(double s[3]);
